Saving MaxMurder.R

+++ MaxMurder.R
@@ -4,5 +4,5 @@ library(ggplot2)
 #library(dplyr)
 df<- final_data
 
-df$NumMurder = df$Murder*df$population/100000
+df$NumMurder = int(df$Murder*df$population/100000)
 str(df)
